services/taskservice/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
sqlparse==0.5.5
pymongo==4.6.1
mongoengine==0.29.1
